feat(frontend): update ic-js v82 and wallet signer v2 (#10215)
# Motivation

Today we released ic-js
[v82](https://github.com/dfinity/icp-js-canisters/releases/tag/v82) and
the OISY wallet signer
[v2](https://github.com/dfinity/oisy-wallet-signer/releases/tag/v2.0.0)
which both references the new `@icp-sdk/core` as peer deps. Deps we have
now upgraded in #9192, so we also want to upgrade those library.

# Changes

- `npm run update:ic-js`
- npm rm & i @dfinity/oisy-wallet-signer
- npm rm & i @dfinity/zod-schemas
- Update various imports that were referencing `/dist/` Candid types.
Those imports does not work anymore with newest ic-js. We could
potentially adapt the path but, better use the official re-export types
which is cleaner.