fix(frontend): Optimize spinner (#10237)
# Motivation

After some analysis it turns out the initial spinner is one of the
reasons IOS browsers crash frequently. We optimize it to get rid of
several issues.

The main problem seems to be that we currently just remove the spinner
via direct DOM manipulation without halting the CSS animations, that use
3D transform. When the DOM element is removed the CSS animations can
still linger in the GPU being calculated for an inexisting element.

To address all these issues, several adjustments had to be made:

# Changes

- Simple div instead of SVG
- No transforms except rotate for the animation, especially no 3D
transforms
- Cleanly remove the animations and after request a seperate animation
frame to remove the DOM element
- Limit animation to 40 iterations (30 seconds) to have a fallback when
the DOM element isnt removed as safely as expected

# Tests


https://github.com/user-attachments/assets/3e2787e3-0239-4297-ab54-68d361168ae7

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>