feat(frontend): auth broadcast singleton (#10271)
# Motivation

We don't face this issue in OISY, but as I was releasing Juno to
production yesterday, I encountered a race condition caused by the
authentication broadcast feature. The tab where the new user signed in
emitted a message, which resulted in the authentication store being
updated twice. This ultimately triggered two identical functions, both
attempting to create a wallet.

Technically, this wasn't an issue, but visually it caused errors to
appear for the new user because one call failed while the other
succeeded.

To prevent a similar issue from potentially occurring in OISY, we
discussed implementing the same guard - i.e. making the broadcast
provider ignore self-messages by using IDs to identify emitters.

# Changes

- Transform `AuthBroadcastChannel` into a singleton
- Generate an ID in the constructor
- Emit message and the ID in the `postMessage`
- On message, discard message received from self
- Rename `close` to `destroy` to reset the singleton