feat(frontend): Cache media status for URLs (#10288)
# Motivation

The CORS errors are polluting too much our browser. Unfortunately [they
cannot really be
silenced](https://stackoverflow.com/questions/52807184/how-to-hide-console-status-error-message-while-fetching-in-react-js).

Since it happens only when we use `fetch` to check the media status, we
can try to reduce at least the times we receive the same error for the
same URL.

In the same session/page, we don't really need to reload the media
status several times for the same URL.

# Changes

- Add cache for media status per URL.
- Add NFT util to use the cache, or fetch the media status and cache it.
- Adapt the code.

# Tests

Provided tests.