fix(frontend): exclude fees from receive amount icp swap (#10296)
# Motivation

We encountered an issue where ICP Swap returns the gross receive amount.
To display the correct value to the user, we need to subtract the
destination token fee.

# Changes

Added a check for the net receiveAmount and updated the ICP Swap utility
to return it in the mapping.

# Tests

Added test coverage for the new logic.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>