feat(frontend): Add worker ID to `AppWorker` class (#10327)
# Motivation

We will use the workers as singleton. However, there is no way of
connecting automatically each posted message to their correct listener,
unless we completely rewrite the way we use workers.

Furthermore, not all workers will be singleton. And some of them will be
only for iOS (to try and reduce the Jetsam issue).

An alternative solution is to provide each worker with an unique ID, and
wrap the listener to such ID: in case the worker is a singleton, the
listener will be triggered only for the correct worker ID.

In this PR, we add the worker ID, since it can be independent of the
rest of the flow.
 
# Changes

- Create worker ID on worker initialization.
- Send always the worker ID with the message.

# Tests

Since we are here, we add tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>