chore(frontend): Remove unused toast error in Exchange worker (#10342)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/631, the error toast
of the exchange was hidden because it was causing too many issues.
Ultimately we decided that in any case we will not the toast error for
exchange error, just a console log, to avoid polluting the user
interface.

We already have skeletons and visual feedbacks for errors while fetching
exchange rate.