feat(frontend): use ICRC index-ng transaction types (#10359)
# Motivation

ic-js types have become strictier as we modernized it's bundle and
tsconfig. As a result, we have some kind of mismatch with TypeScript
types definition usage between Icrc Index and Index-NG canisters. As a
result, trying to migrate to `@icp-sdk/canisters` fails (see #10358).

I don't remember if it was in OISY or one of my project, but I reckon we
recently had to adapt types in a worker. This PR is similar. It uses the
types of the Index NG which by definition is correct, since OISY is
using the Index NG since a long time.

# Changes

- Update `IcrcTransaction` and `IcrcTransactionWithId` to index NG
canister

Signed-off-by: David Dal Busco <david.dalbusco@outlook.com>