build(frontend): migrate dependencies to @icp-sdk/canisters v1 (#10358)
# Motivation

We want to start using `@icp-sdk/canisters` in OISY. Migrating is
painless, as the multi-entry libraries re-export the legacy libraries as
peers. That’s why we can bump the libs without modifying the imports.

# Notes

We are migrating to `@icp-sdk/canisters@1.1.1` first because the second
version introduces breaking changes in the Candid IDL declarations which
needs several adjustements including in the tooling. See draft #10355.

# Changes

- Replace `@dfinity/...` with `@icp-sdk/canisters`
- Update `npm run...` scripts accordingly
- Add `zod-schemas` to the ic-js scripts - not sure it's required but
recently we had a conflict in a project when we released the major
version

Signed-off-by: David Dal Busco <david.dalbusco@outlook.com>