feat(frontend): Make all IC wallet worker singletons in iOS (#10390)
# Motivation

We are facing serious _jetsam_ problems with iOS (mostly iPhones).
Basically, the memory pressure is such that, on some wallets, the OISY
page auto-refreshes or crashes at first touches, making it almost
impossible to use it.

We are trying to make it "lighter". One way to do it is to make the IC
wallet workers as singletons for iOS.