refactor(frontend): Avoid referring to subfolder for `icp_swap_factory` bindings (#10436)
# Motivation

With the new bindings generation script (PR
https://github.com/dfinity/oisy-wallet/pull/10409), we prefer to not
have the subfolder `declarations` that bindgen creates, but to
consolidate the bindings in the same folder.

However, our current imports are using such subfolder.

The plan is:

- Undo the removal of the subfolder (temporarily).
- Adapt all the imports to use the "old" main folder for declarations
(split in a few PRs, since it is quite a lot of changes).
- Re-remove the subfolder.


This PR is part of the second step: we undo the reference to subfolder
for the `icp_swap_factory` bindings that was done in PR
https://github.com/dfinity/oisy-wallet/pull/9566