feat(did): Reinstate the removal of subfolder `declarations` (#10444)
# Motivation

With the new bindings generation script (PR
https://github.com/dfinity/oisy-wallet/pull/10409), we prefer to not
have the subfolder `declarations` that bindgen creates, but to
consolidate the bindings in the same folder.

However, our current imports are using such subfolder.

The plan is:

- Undo the removal of the subfolder (temporarily).
- Adapt all the imports to use the "old" main folder for declarations
(split in a few PRs, since it is quite a lot of changes).
- Re-remove the subfolder.


This PR is the last step: we reinstate the removal of the subfolder
`declarations`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>