feat(backend)!: Add EXT v2 tokens to Custom tokens variants (#10360)
# Motivation

We are going to onboard NFTs in IC. One of the most used standard is
[EXT v2 Token](https://github.com/Toniq-Labs/ext-v2-token).

We need to include it among the variants for custom tokens, since we are
not going to treat them the same as ICRC tokens.

BREAKING CHANGE: The interface of the backend now includes a new
variant.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>