fix(frontend): Use same list of tokens for Manage tokens and Filter tokens (#10456)
# Motivation

The filter in the token list uses a different base list of tokens from
the one of the manage tokens modal.

In fact, it is currently not considering all ICRC tokens: SNSes are
defined as environment variables. These tokens are not necessarily
loaded at boot time if the user has not added them to their list of
custom tokens.

So, we need to make them use the same list, since they have the same
scope more or less.

# Changes

- Create new derived store with all fungible tokens (all of them
including SNSes).
- Create new derived store for all fungible tokens filtered per network.
- Use the last one in the filter of the tokens list.
- Remove unused derived store `fungibleNetworkTokens`.

# Tests

Created tests.