feat(build): Create script to load the NFT EXT v2 collections (#10497)
# Motivation

There is no direct on-chain way to know the list of existing EXT v2
tokens/collections. But we require it, since we will create the same
type of worker that we currently have for Ethereum/EVM NFTs: we want to
periodically check if the user has any NFT pertaining to a list of
collections.

So, as first step we need to maintain such a list (very similar to what
we do for SNSes).

For now we found two sources for a extensive list:

- [Toniq exchange](https://toniq.io/);
- [nftGeek](https://t5t44-naaaa-aaaah-qcutq-cai.raw.icp0.io/).

Scraping their network requests, we find the URL to be queried.

# Note

The user will have the possibility to manually add a collection if it is
not in the maintained list yet.

# Changes

- Create small script to fetch and parse data from Toniq.
- Same for nftGeek.
- Merge the data and save them: we will not remove the existing
collections that are not in the fetched data anymore.
- Create script to run the module.

# Tests

I manually ran the script and save the created file.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>