refactor(frontend): Move ETH NFT service to its own module (#10538)
# Motivation

We are going to have NFT services for EXT tokens too. So, to keep
everything more organized, we are going to split the methods
per-network. In this PR, we move the Ethereum/EVM NFT service
`loadNftsByNetwork` to its own module inside the `eth` folder.

Since we are here, we add tests.