feat(frontend): Fetch EXT NFT collections in `LoaderCollections` (#10530)
# Motivation

We want to load actively the EXT collections that the user owns NFTs in.
The flow is similar to ERC collections.

However, for now, we have no other way of knowing what EXT
tokens/collections the user has NFT with.

So, we loop through a curated list, and we use the method
`getTokensByOwner` to check if the user has any NFTs from that
collection.