feat(frontend): NFT media status should be compulsory (#10600)
# Motivation

The media status of an NFT is used to display the image itself or a
placeholder (for example for invalid data). Since we consider the lack
of an image URL as invalid data, we should have ALWAYS the media status
fore each NFT, independently from the URL.

# Changes

- Make field `mediaStatus` compulsory in `NftSchema`.
- Adapt mapper for EXT NFTs to add the field `mediaStatus`: we re-use
the service `getMediaStatusOrCache`, like for ERC NFTs.
- Adapt the consumers of the mapper.

# Tests

Adapted tests.