feat(frontend): Load EXT NFT collections only on request (#10622)
# Motivation

The EXT NFT collections are refreshed based on a curated list of EXT
tokens. However, since they are a lot, it does not really make sense to
keep refreshing them: it will only slow down the UI.

So, we will refresh it on demand.

As first step, we adapt component `LoaderCollections` to accept
triggered calls.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>