feat(frontend): minimal GLDT unstake fraction (#10682)
# Motivation

We need to fix the case where user inputs a very low amount of GLDT to
unstake which results in fraction being 0 and hence gldt-stake canister
error. This approach is needed only until Gold DAO team provides us with
the endpoint that accepts amount to unstake as a number and not as a
percentage.