feat(frontend): use @icp-sdk/canisters with embedded ic-mgmt (#10698)
# Motivation

We want to embed the code in `@icp-sdk/canisters` instead of
re-exporting peer dependencies on legacy libraries or, conversely,
re-exporting `@icp-sdk/canisters` as a peer from the legacy
`@dfinity/...` dependencies.

We also want to use this opportunity to migrate the types to a DID and
IDL namespace approach and clean-up the naming inconsistencies (camel
case vs upper case).

We are going to make one major release, but we'll integrate the work
step by step, starting with the IC mgmt.

# Notes

This PR also contains the CMC embeded but, it's unused in OISY:

# Changes

- Upgrade `@icp-sdk/canisters@next`
- Remove inexisting enum and replace with object 
- Update canister name

# Unrelated Changes

In this week Candid new types, there is a new field `fee_collector` in
`@dfinity/ledger-icrc`. Unfortunatelly I couldn't - I tried - update it
separatly. I would had need to use an old version of next and then just
update the lib. Felt like lot of work to just add the field to two test
mocks. Therefore, I provide this change in this PR as well.

---------

Signed-off-by: David Dal Busco <david.dalbusco@outlook.com>