feat(frontend): Component to review EXT token when custom-adding (#10747)
# Motivation

We want to add custom tokens for EXT standard too.

Copying directly from `IcAddIcrcTokenReview`, we create the parallel
component for the EXT tokens: `IcAddExtTokenReview`.

# Changes

- Extend the list of add-token data with EXT tokens: the only thing
required is the canister ID.
- Copy `IcAddIcrcTokenReview` in `IcAddExtTokenReview` and remove the
logic for the Index Canister, and the metadata: for now we don't really
have metadata for EXT tokens.
- Adapt `AddTokenReviewByNetwork` using the same logic of the NFT page.

# Tests

Current tests should work.