feat(frontend): Allow to reset custom tokens via identifier (#10755)
# Motivation

We want to generalize how we reset the custom token store, to make it
more scalable. So, we first need another method to reset the store when
we still don't know the token ID (for example, when we are saving new
custom tokens).

# Changes

- Add new method `resetByIdentifier` to custom token store: it use
everything apart from token ID.
- Fix the correct types for `Identifier` of custom tokens.
- Apply the new method to service `saveCustomTokens`.

# Tests

Added tests.