feat(frontend): update ic-js next and use did namespace for ckBtc and ckEth (#10783)
# Motivation

We want to ship `@icp-sdk/canisters` without peer dependencies - i.e. we
want to move and embed the code in this library.
We also want to start using a "did namespace" pattern".

This PRs bumps the libraries to next which was upgraded for ckBtc and
ckEth and also update the imports accordingly.

# Changes

- Upgrade next
- Adapt imports for DID namespace
- Adapt imports for canisters using camel case (e.g. CkBTC -> CkBtc)