fix(frontend): Resolve plausible imports and related tests (#10801)
# Motivation

When I bumped `vitest to v4`, I changed the dynamic import or package
`@plausible-analytics/tracker` to use a constant, since it was raising
issues with the tests:

> Error: Failed to resolve entry for package
"@plausible-analytics/tracker". The package may have incorrect
main/module/exports specified in its package.json.
>   Plugin: vite:import-analysis
> File:
/Users/antonio.ventilii/projects/oisy-wallet/src/frontend/src/lib/services/analytics-wrapper.ts:1:61
> 1 | export const loadPlausibleTracker = async () => await
import("@plausible-analytics/tracker");
> | ^
> 

However, Plausible detected that we had an incorrect usage of the
package at this point and the events are not being registered anymore.

To fix it, we revert the dynamic import and, to fix the tests, we mock
the plausible imports.

# Changes

- Revert the dynamic import of Plausible tracker to use a string
directly.
- Create mock class for the tracker.
- Resolve the mock tracker as alias in vitest.

# Tests

Current tests should work.