feat(frontend): use ic-js next with DID namespace for ledgers, Nns and Sns (#10806)
# Motivation

We want to move the source of of the ic-js legacy libraries to the
multi-entries library `@icp-sdk/canisters`.
At the same time, we also want to bring consistency and start using a
pattern we called "DID namespace usage".

Unfortunatelly, because of interdepency between ledgers used in Nns and
Sns libs, we had to pack everything at once.

This PRs bumps the latest next libs and adapt the interfaces.

NO functional changes.

# Changes

- `npm run upgrade:next`
- Update imports and usage for DID namespace