feat(devops): Cancel CI to update test coverage thresholds in progress (#10861)
# Motivation

There is a small issue with not cancelling a CI to update the test
coverage thresholds that is in progress: if there is another PR that
updated the thresholds, and that was merged while the CI is running, the
new PR will not know of those updated values.

So, when we rebase it, we risk to create an empty PR, and we need to
close it.