feat(frontend): Hide Scanner button temporarily (#10925)
# Motivation

On mobile, the header has too many buttons for now.

We are planning on integrating WalletConnect into the Scanner, but for
now, the Scanner only has the Pay function, that is accessible via the
Pay button too.

So, we can temporarily remove the Scanner button, awaiting the complete
integration of WalletConnect.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>