refactor(frontend): Rename transfers methods of EXT canister (#10937)
# Motivation

Endpoints `transfer` and `ext_transfer` of EXT token canister are
actually equivalent, with the latter being an alias of the first one.

However, not all of them implement `ext_transfer`, so we can actually
always try `transfer` first and fall back on the alias for redundancy.

To this scope, it does not make sense to name them with `legacy` suffix,
but only with alias suffix.