build(frontend): temporarily revert update agentjs @icp-sdk/core and auth to v5 (#10944)
# Motivation

In PR https://github.com/dfinity/oisy-wallet/pull/10933 we bumped
agentjs, @icp-sdk/core, and auth to v5 (plus related libraries for
compatibility, like OISY Wallet Signer).

We tested our Signer before going live and didn't have any issues - all
good from our perspective.

However, some user is experiencing issues with the Signer.

So, it's likely because something isn't compatible when the client uses
signer-js with Agent-JS v4.

This should be reported to the SDK team if that's the case -
incompatibility between client in v4 and signer in v5.
Or signer-js can be upgraded to v5 but, then it requires all clients to
upgrade it.

As temporarily patch in OISY, we roll back ic-js and signer to the
previous versions (rolling back PR
https://github.com/dfinity/oisy-wallet/pull/10933).

Credits to @peterpeterparker to finding it out.